+++ src/mol-canvas3d/helper/interaction-events.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2018-2023 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ * Copyright (c) 2018-2025 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author David Sehnal <david.sehnal@gmail.com>
  * @author Alexander Rose <alexander.rose@weirdbyte.de>
@@ -14,13 +14,13 @@ import { Camera } from '../camera';
 import { ParamDefinition as PD } from '../../mol-util/param-definition';
 import { Bond } from '../../mol-model/structure';
 import { TrackballControls } from '../controls/trackball';
+import { Ray3D } from '../../mol-math/geometry/primitives/ray3d';
+import { AsyncPickData } from '../passes/pick';
 
 type Canvas3D = import('../canvas3d').Canvas3D
 type HoverEvent = import('../canvas3d').Canvas3D.HoverEvent
 type DragEvent = import('../canvas3d').Canvas3D.DragEvent
 type ClickEvent = import('../canvas3d').Canvas3D.ClickEvent
-
-enum InputEvent { Move, Click, Drag }
 
 const tmpPosA = Vec3();
 const tmpPos = Vec3();
@@ -29,6 +29,7 @@ const tmpNorm = Vec3();
 export const Canvas3dInteractionHelperParams = {
     maxFps: PD.Numeric(30, { min: 10, max: 60, step: 10 }),
     preferAtomPixelPadding: PD.Numeric(3, { min: 0, max: 20, step: 1 }, { description: 'Number of extra pixels at which to prefer atoms over bonds.' }),
+    convertCoordsToRay: PD.Boolean(false, { description: 'Convert screen coordinates to ray for picking.' }),
 };
 export type Canvas3dInteractionHelperParams = typeof Canvas3dInteractionHelperParams
 export type Canvas3dInteractionHelperProps = PD.Values<Canvas3dInteractionHelperParams>
@@ -47,10 +48,11 @@ export class Canvas3dInteractionHelper {
     private endX = -1;
     private endY = -1;
 
-    private id: PickingId | undefined = void 0;
+    private ray: Ray3D | undefined = void 0;
+
+    private pickData: AsyncPickData | undefined = void 0;
     private position: Vec3 | undefined = void 0;
 
-    private currentIdentifyT = 0;
     private isInteracting = false;
 
     private prevLoci: Representation.Loci = Representation.Loci.Empty;
@@ -68,46 +70,67 @@ export class Canvas3dInteractionHelper {
         Object.assign(this.props, props);
     }
 
-    private identify(e: InputEvent, t: number) {
+    private getTarget(): Vec2 | Ray3D {
+        if (this.ray) {
+            return this.ray;
+        } else if (this.props.convertCoordsToRay) {
+            return this.camera.getRay(Ray3D(), this.endX, this.input.height - this.endY);
+        } else {
+            return Vec2.create(this.endX, this.endY);
+        }
+    }
+
+    private handleMove() {
         const xyChanged = this.startX !== this.endX || this.startY !== this.endY || (this.input.pointerLock && !this.controls.isMoving);
 
-        if (e === InputEvent.Drag) {
-            if (xyChanged && !this.outsideViewport(this.startX, this.startY)) {
-                this.events.drag.next({ current: this.prevLoci, buttons: this.buttons, button: this.button, modifiers: this.modifiers, pageStart: Vec2.create(this.startX, this.startY), pageEnd: Vec2.create(this.endX, this.endY) });
-
-                this.startX = this.endX;
-                this.startY = this.endY;
-            }
-            return;
-        }
-
         if (xyChanged) {
-            const pickData = this.canvasIdentify(this.endX, this.endY);
-            this.id = pickData?.id;
-            this.position = pickData?.position;
+            const pickData = this.canvasAsyncIdentify(this.getTarget());
+            this.pickData = pickData;
             this.startX = this.endX;
             this.startY = this.endY;
         }
+    }
 
-        if (e === InputEvent.Click) {
-            const loci = this.getLoci(this.id, this.position);
-            this.events.click.next({ current: loci, buttons: this.buttons, button: this.button, modifiers: this.modifiers, page: Vec2.create(this.endX, this.endY), position: this.position });
-            this.prevLoci = loci;
-            return;
-        }
-
-        if (!this.inside || this.currentIdentifyT !== t || !xyChanged || this.outsideViewport(this.endX, this.endY)) return;
-
-        const loci = this.getLoci(this.id, this.position);
-        this.events.hover.next({ current: loci, buttons: this.buttons, button: this.button, modifiers: this.modifiers, page: Vec2.create(this.endX, this.endY), position: this.position });
+    private handleClick() {
+        const pickData = this.canvasIdentify(this.getTarget());
+        const loci = this.getLoci(pickData?.id, pickData?.position);
+        this.events.click.next({ current: loci, buttons: this.buttons, button: this.button, modifiers: this.modifiers, page: Vec2.create(this.endX, this.endY), position: pickData?.position });
         this.prevLoci = loci;
     }
 
+    private handleDrag() {
+        const xyChanged = this.startX !== this.endX || this.startY !== this.endY || (this.input.pointerLock && !this.controls.isMoving);
+
+        if (xyChanged && !this.outsideViewport(this.startX, this.startY, this.ray)) {
+            this.events.drag.next({ current: this.prevLoci, buttons: this.buttons, button: this.button, modifiers: this.modifiers, pageStart: Vec2.create(this.startX, this.startY), pageEnd: Vec2.create(this.endX, this.endY) });
+
+            this.startX = this.endX;
+            this.startY = this.endY;
+        }
+    }
+
     tick(t: number) {
-        if (this.inside && t - this.prevT > 1000 / this.props.maxFps) {
-            this.prevT = t;
-            this.currentIdentifyT = t;
-            this.identify(this.isInteracting ? InputEvent.Drag : InputEvent.Move, t);
+        if (this.pickData) {
+            if (this.pickData.check()) {
+                const pickData = this.pickData.get();
+                this.position = pickData?.position;
+                if (this.inside) {
+                    const loci = this.getLoci(pickData?.id, pickData?.position);
+                    this.events.hover.next({ current: loci, buttons: this.buttons, button: this.button, modifiers: this.modifiers, page: Vec2.create(this.endX, this.endY), position: pickData?.position });
+                    this.prevLoci = loci;
+                }
+
+                this.pickData = undefined;
+            }
+        } else {
+            if (this.inside && t - this.prevT > 1000 / this.props.maxFps) {
+                this.prevT = t;
+                if (this.isInteracting) {
+                    this.handleDrag();
+                } else {
+                    this.handleMove();
+                }
+            }
         }
     }
 
@@ -119,22 +142,24 @@ export class Canvas3dInteractionHelper {
         }
     }
 
-    private move(x: number, y: number, buttons: ButtonsType, button: ButtonsType.Flag, modifiers: ModifiersKeys) {
+    private move(x: number, y: number, buttons: ButtonsType, button: ButtonsType.Flag, modifiers: ModifiersKeys, ray?: Ray3D) {
         this.inside = true;
         this.buttons = buttons;
         this.button = button;
         this.modifiers = modifiers;
+        this.ray = ray;
         this.endX = x;
         this.endY = y;
     }
 
-    private click(x: number, y: number, buttons: ButtonsType, button: ButtonsType.Flag, modifiers: ModifiersKeys) {
+    private click(x: number, y: number, buttons: ButtonsType, button: ButtonsType.Flag, modifiers: ModifiersKeys, ray?: Ray3D) {
         this.endX = x;
         this.endY = y;
         this.buttons = buttons;
         this.button = button;
         this.modifiers = modifiers;
-        this.identify(InputEvent.Click, 0);
+        this.ray = ray;
+        this.handleClick();
     }
 
     private drag(x: number, y: number, buttons: ButtonsType, button: ButtonsType.Flag, modifiers: ModifiersKeys) {
@@ -143,7 +168,7 @@ export class Canvas3dInteractionHelper {
         this.buttons = buttons;
         this.button = button;
         this.modifiers = modifiers;
-        this.identify(InputEvent.Drag, 0);
+        this.handleDrag();
     }
 
     private modify(modifiers: ModifiersKeys) {
@@ -152,7 +177,9 @@ export class Canvas3dInteractionHelper {
         this.events.hover.next({ current: this.prevLoci, buttons: this.buttons, button: this.button, modifiers: this.modifiers, page: Vec2.create(this.endX, this.endY), position: this.position });
     }
 
-    private outsideViewport(x: number, y: number) {
+    private outsideViewport(x: number, y: number, ray?: Ray3D) {
+        if (ray) return false;
+
         const { input, camera: { viewport } } = this;
         x *= input.pixelRatio;
         y *= input.pixelRatio;
@@ -189,7 +216,7 @@ export class Canvas3dInteractionHelper {
         this.ev.dispose();
     }
 
-    constructor(private canvasIdentify: Canvas3D['identify'], private lociGetter: Canvas3D['getLoci'], private input: InputObserver, private camera: Camera, private controls: TrackballControls, props: Partial<Canvas3dInteractionHelperProps> = {}) {
+    constructor(private canvasIdentify: Canvas3D['identify'], private canvasAsyncIdentify: Canvas3D['asyncIdentify'], private lociGetter: Canvas3D['getLoci'], private input: InputObserver, private camera: Camera, private controls: TrackballControls, props: Partial<Canvas3dInteractionHelperProps> = {}) {
         this.props = { ...PD.getDefaultValues(Canvas3dInteractionHelperParams), ...props };
 
         input.drag.subscribe(({ x, y, buttons, button, modifiers }) => {
@@ -198,14 +225,14 @@ export class Canvas3dInteractionHelper {
             this.drag(x, y, buttons, button, modifiers);
         });
 
-        input.move.subscribe(({ x, y, inside, buttons, button, modifiers, onElement }) => {
+        input.move.subscribe(({ x, y, inside, buttons, button, modifiers, onElement, ray }) => {
             if (!inside || this.isInteracting) return;
             if (!onElement) {
                 this.leave();
                 return;
             }
             // console.log('move');
-            this.move(x, y, buttons, button, modifiers);
+            this.move(x, y, buttons, button, modifiers, ray);
         });
 
         input.leave.subscribe(() => {
@@ -213,10 +240,10 @@ export class Canvas3dInteractionHelper {
             this.leave();
         });
 
-        input.click.subscribe(({ x, y, buttons, button, modifiers }) => {
-            if (this.outsideViewport(x, y)) return;
+        input.click.subscribe(({ x, y, buttons, button, modifiers, ray }) => {
+            if (this.outsideViewport(x, y, ray)) return;
             // console.log('click');
-            this.click(x, y, buttons, button, modifiers);
+            this.click(x, y, buttons, button, modifiers, ray);
         });
 
         input.interactionEnd.subscribe(() => {